Further transition to uint64_t variables being used for array sizes in packets.
The lightweight version of test_arrays was not updated for this change, nor
was the code that handles topology updates on the lightweight side. All
MRNet tests now work with lightweight backends when build is configured with
--enable-debug.  There is still a problem when debug is off for lightweight.

+++ src/lightweight/FilterDefinitions.c
@@ -41,14 +41,14 @@ void tfilter_TopoUpdate(vector_t * ipackets,
     uint32_t * rcrank_arr = NULL;
     char ** rchost_arr = NULL;
     uint16_t * rcport_arr = NULL;
-    size_t rarray_len = 0;
+    uint64_t rarray_len = 0;
 
     int * type_arr = NULL;
     uint32_t * prank_arr = NULL;
     uint32_t * crank_arr = NULL;
     char ** chost_arr = NULL;
     uint16_t * cport_arr = NULL;
-    unsigned arr_len = 0;
+    uint64_t arr_len = 0;
     unsigned long arr_len_long = 0;
 
     char * format_string = NULL;
